+++ src/Contracts/Permission.php
@@ -6,22 +6,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 interface Permission
 {
-    /**
-     * A permission can be applied to roles.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
-     */
     public function roles(): BelongsToMany;
 
-    /**
-     * Find a permission by its name.
-     *
-     * @param string $name
-     * @param string|null $guardName
-     *
-     * @throws \Konekt\Acl\Exceptions\PermissionDoesNotExist
-     *
-     * @return Permission
-     */
-    public static function findByName(string $name, $guardName): self;
+    public static function findByName(string $name, ?string $guardName = null): ?self;
 }