+++ src/Contracts/Role.php
@@ -7,42 +7,11 @@ use Konekt\Acl\Exceptions\RoleDoesNotExist;
 
 interface Role
 {
-    /**
-     * A role may be given various permissions.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
-     */
     public function permissions(): BelongsToMany;
 
-    /**
-     * Find a role by its name and guard name.
-     *
-     * @param string $name
-     * @param string|null $guardName
-     *
-     * @return \Konekt\Acl\Contracts\Role
-     *
-     * @throws RoleDoesNotExist
-     */
-    public static function findByName(string $name, $guardName): self ;
+    public static function findByName(string $name, ?string $guardName = null): ?self ;
 
-    /**
-     * Find a role by its id and guard name.
-     * @param int $id
-     * @param string|null $guardName
-     *
-     * @return \Konekt\Acl\Contracts\Role
-     *
-     * @throws \Konekt\Acl\Exceptions\RoleDoesNotExist
-     */
-    public static function findById(int $id, $guardName): self;
+    public static function findById(int $id, ?string $guardName = null): ?self;
 
-    /**
-     * Determine if the user may perform the given permission.
-     *
-     * @param string|Permission $permission
-     *
-     * @return bool
-     */
-    public function hasPermissionTo($permission): bool;
+    public function hasPermissionTo(string|Permission $permission): bool;
 }