+++ tests/HasPermissionsTest.php
@@ -58,44 +58,12 @@ class HasPermissionsTest extends TestCase
     {
         $user1 = User::create(['email' => 'user1@test.com']);
         $user2 = User::create(['email' => 'user2@test.com']);
-        $user1->givePermissionTo(['edit-articles', 'edit-news']);
+        $user1->givePermissionTo('edit-articles', 'edit-news');
         $this->testUserRole->givePermissionTo('edit-articles');
         $user2->assignRole('testRole');
 
-        $scopedUsers1 = User::permission('edit-articles')->get();
-        $scopedUsers2 = User::permission(['edit-news'])->get();
-
-        $this->assertEquals($scopedUsers1->count(), 2);
-        $this->assertEquals($scopedUsers2->count(), 1);
-    }
-
-    /** @test */
-    public function it_can_scope_users_using_an_array()
-    {
-        $user1 = User::create(['email' => 'user1@test.com']);
-        $user2 = User::create(['email' => 'user2@test.com']);
-        $user1->givePermissionTo(['edit-articles', 'edit-news']);
-        $this->testUserRole->givePermissionTo('edit-articles');
-        $user2->assignRole('testRole');
-
-        $scopedUsers1 = User::permission(['edit-articles', 'edit-news'])->get();
-        $scopedUsers2 = User::permission(['edit-news'])->get();
-
-        $this->assertEquals($scopedUsers1->count(), 2);
-        $this->assertEquals($scopedUsers2->count(), 1);
-    }
-
-    /** @test */
-    public function it_can_scope_users_using_a_collection()
-    {
-        $user1 = User::create(['email' => 'user1@test.com']);
-        $user2 = User::create(['email' => 'user2@test.com']);
-        $user1->givePermissionTo(['edit-articles', 'edit-news']);
-        $this->testUserRole->givePermissionTo('edit-articles');
-        $user2->assignRole('testRole');
-
-        $scopedUsers1 = User::permission(collect(['edit-articles', 'edit-news']))->get();
-        $scopedUsers2 = User::permission(collect(['edit-news']))->get();
+        $scopedUsers1 = User::havingPermission('edit-articles')->get();
+        $scopedUsers2 = User::havingPermission('edit-news')->get();
 
         $this->assertEquals($scopedUsers1->count(), 2);
         $this->assertEquals($scopedUsers2->count(), 1);
@@ -107,13 +75,9 @@ class HasPermissionsTest extends TestCase
         $user1 = User::create(['email' => 'user1@test.com']);
         $user1->givePermissionTo($this->testUserPermission->name);
 
-        $scopedUsers1 = User::permission($this->testUserPermission)->get();
-        $scopedUsers2 = User::permission([$this->testUserPermission])->get();
-        $scopedUsers3 = User::permission(collect([$this->testUserPermission]))->get();
+        $scopedUsers1 = User::havingPermission($this->testUserPermission)->get();
 
         $this->assertEquals($scopedUsers1->count(), 1);
-        $this->assertEquals($scopedUsers2->count(), 1);
-        $this->assertEquals($scopedUsers3->count(), 1);
     }
 
     /** @test */
@@ -125,7 +89,7 @@ class HasPermissionsTest extends TestCase
         $user1->assignRole('testRole');
         $user2->assignRole('testRole');
 
-        $scopedUsers = User::permission('edit-articles')->get();
+        $scopedUsers = User::havingPermission('edit-articles')->get();
 
         $this->assertEquals($scopedUsers->count(), 2);
     }
@@ -135,10 +99,10 @@ class HasPermissionsTest extends TestCase
     {
         $user1 = User::create(['email' => 'user1@test.com']);
         $user2 = User::create(['email' => 'user2@test.com']);
-        $user1->givePermissionTo(['edit-news']);
-        $user2->givePermissionTo(['edit-articles', 'edit-news']);
+        $user1->givePermissionTo('edit-news');
+        $user2->givePermissionTo('edit-articles', 'edit-news');
 
-        $scopedUsers = User::permission('edit-news')->get();
+        $scopedUsers = User::havingPermission('edit-news')->get();
 
         $this->assertEquals($scopedUsers->count(), 2);
     }
@@ -148,18 +112,18 @@ class HasPermissionsTest extends TestCase
     {
         $this->expectException(PermissionDoesNotExist::class);
 
-        User::permission('testAdminPermission')->get();
+        User::havingPermission('testAdminPermission')->get();
 
         $this->expectException(GuardDoesNotMatch::class);
 
-        User::permission($this->testAdminPermission)->get();
+        User::havingPermission($this->testAdminPermission)->get();
     }
 
     /** @test */
     public function it_doesnt_detach_permissions_when_soft_deleting()
     {
         $user = SoftDeletingUser::create(['email' => 'test@example.com']);
-        $user->givePermissionTo(['edit-news']);
+        $user->givePermissionTo('edit-news');
         $user->delete();
 
         $user = SoftDeletingUser::withTrashed()->find($user->id);
@@ -170,7 +134,7 @@ class HasPermissionsTest extends TestCase
     /** @test */
     public function it_can_give_and_revoke_multiple_permissions()
     {
-        $this->testUserRole->givePermissionTo(['edit-articles', 'edit-news']);
+        $this->testUserRole->givePermissionTo('edit-articles', 'edit-news');
 
         $this->assertEquals(2, $this->testUserRole->permissions()->count());